Ten

Description

This module serves as the main entry point for the Ten library, re-exporting the core functionality while providing a simplified API.

Core types

module Ten.Core

Database operations

module Ten.DB.Core

Client functionality

connectToTenDaemon :: FilePath -> UserCredentials -> IO (Either BuildError (DaemonConnection 'Builder))

Connect to a Ten daemon

disconnectFromDaemon :: DaemonConnection 'Builder -> IO ()

Disconnect from a Ten daemon

runClientCommand :: DaemonConnection 'Builder -> (DaemonConnection 'Builder -> IO (Either BuildError a)) -> IO (Either BuildError a)

Run a client command against a daemon

Version information

version :: (Int, Int, Int)

Version information based on protocol version

versionString :: String

Version string for display purposes