Ten.DB.References

Documentation

class CanRegisterReferences (t :: PrivilegeTier) where

Type class for reference registration operations

Methods

registerReferenceImpl :: forall (p :: Phase). Database t -> StorePath -> StorePath -> TenM p t ()

Register a single reference between two store paths

registerReferencesImpl :: forall (p :: Phase). Database t -> StorePath -> Set StorePath -> TenM p t ()

Register multiple references from one referrer

registerPathReferencesImpl :: forall (p :: Phase). Database t -> FilePath -> StorePath -> TenM p t Int

Register references for a path after scanning the file

class CanQueryReferences (t :: PrivilegeTier) where

Type class for reference query operations

Methods

getDirectReferencesImpl :: forall (p :: Phase). Database t -> StorePath -> TenM p t (Set StorePath)

Get direct references from a store path

getReferencesImpl :: forall (p :: Phase). Database t -> StorePath -> TenM p t (Set StorePath)

Get all references from a path (direct and indirect)

getDirectReferrersImpl :: forall (p :: Phase). Database t -> StorePath -> TenM p t (Set StorePath)

Get direct referrers to a store path

getReferrersImpl :: forall (p :: Phase). Database t -> StorePath -> TenM p t (Set StorePath)

Get all referrers to a path (direct and indirect)

class CanManageReachability (t :: PrivilegeTier) where

Type class for reachability operations

Methods

computeReachablePathsImpl :: forall (p :: Phase). Database t -> Set StorePath -> TenM p t (Set StorePath)

Compute all paths reachable from a set of roots

findPathsClosureImpl :: forall (p :: Phase). Database t -> Set StorePath -> TenM p t (Set StorePath)

Find the transitive closure of paths

findPathsClosureWithLimitImpl :: forall (p :: Phase). Database t -> Set StorePath -> Int -> TenM p t (Set StorePath)

Find the transitive closure of paths with depth limit

isPathReachableImpl :: forall (p :: Phase). Database t -> Set StorePath -> StorePath -> TenM p t Bool

Check if a path is reachable from any of the roots

class CanManageGCRoots (t :: PrivilegeTier) where

Type class for GC root operations

Methods

findGCRootsImpl :: forall (p :: Phase). Database t -> FilePath -> TenM p t (Set StorePath)

Find all GC roots

getRegisteredRootsImpl :: forall (p :: Phase). Database t -> TenM p t (Set StorePath)

Get registered roots from the database

class CanScanReferences (t :: PrivilegeTier) where

Type class for reference scanning operations

Methods

scanFileForReferencesImpl :: forall (p :: Phase). FilePath -> FilePath -> TenM p t (Set StorePath)

Scan a file for references to store paths

scanStoreForReferencesImpl :: forall (p :: Phase). FilePath -> TenM p t (Set StorePathReference)

Scan the entire store for references between paths

class CanManageValidPaths (t :: PrivilegeTier) where

Type class for path validity management

Methods

markPathsAsValidImpl :: forall (p :: Phase). Database t -> Set StorePath -> TenM p t ()

Mark a set of paths as valid in the ValidPaths table

markPathsAsInvalidImpl :: forall (p :: Phase). Database t -> Set StorePath -> TenM p t ()

Mark a set of paths as invalid in the ValidPaths table

class CanManageDatabase (t :: PrivilegeTier) where

Type class for database management operations

Methods

vacuumReferenceDbImpl :: forall (p :: Phase). Database t -> TenM p t ()

Vacuum the reference database to optimize performance

validateReferenceDbImpl :: forall (p :: Phase). Database t -> FilePath -> TenM p t (Int, Int)

Validate and repair the reference database

cleanupDanglingReferencesImpl :: forall (p :: Phase). Database t -> TenM p t Int

Cleanup dangling references

registerReference :: forall (t :: PrivilegeTier) (p :: Phase). CanRegisterReferences t => Database t -> StorePath -> StorePath -> TenM p t ()

Context-aware reference registration

registerReferences :: forall (t :: PrivilegeTier) (p :: Phase). CanRegisterReferences t => Database t -> StorePath -> Set StorePath -> TenM p t ()

Context-aware reference registration (multiple)

registerPathReferences :: forall (t :: PrivilegeTier) (p :: Phase). CanRegisterReferences t => Database t -> FilePath -> StorePath -> TenM p t Int

Context-aware path reference registration

getReferences :: forall (t :: PrivilegeTier) (p :: Phase). CanQueryReferences t => Database t -> StorePath -> TenM p t (Set StorePath)

Context-aware full reference query

getReferrers :: forall (t :: PrivilegeTier) (p :: Phase). CanQueryReferences t => Database t -> StorePath -> TenM p t (Set StorePath)

Context-aware full referrer query

getDirectReferences :: forall (t :: PrivilegeTier) (p :: Phase). CanQueryReferences t => Database t -> StorePath -> TenM p t (Set StorePath)

Context-aware direct reference query

getDirectReferrers :: forall (t :: PrivilegeTier) (p :: Phase). CanQueryReferences t => Database t -> StorePath -> TenM p t (Set StorePath)

Context-aware direct referrer query

computeReachablePathsFromRoots :: forall (t :: PrivilegeTier) (p :: Phase). CanManageReachability t => Database t -> Set StorePath -> TenM p t (Set StorePath)

Context-aware compute reachable paths

findPathsClosure :: forall (t :: PrivilegeTier) (p :: Phase). CanManageReachability t => Database t -> Set StorePath -> TenM p t (Set StorePath)

Context-aware path closure

findPathsClosureWithLimit :: forall (t :: PrivilegeTier) (p :: Phase). CanManageReachability t => Database t -> Set StorePath -> Int -> TenM p t (Set StorePath)

Context-aware path closure with limit

isPathReachable :: forall (t :: PrivilegeTier) (p :: Phase). CanManageReachability t => Database t -> Set StorePath -> StorePath -> TenM p t Bool

Context-aware path reachability check

findGCRoots :: forall (t :: PrivilegeTier) (p :: Phase). CanManageGCRoots t => Database t -> FilePath -> TenM p t (Set StorePath)

Context-aware GC roots finder

getRegisteredRoots :: forall (t :: PrivilegeTier) (p :: Phase). CanManageGCRoots t => Database t -> TenM p t (Set StorePath)

Context-aware get registered roots

markPathsAsValid :: forall (t :: PrivilegeTier) (p :: Phase). CanManageValidPaths t => Database t -> Set StorePath -> TenM p t ()

Context-aware mark paths as valid

markPathsAsInvalid :: forall (t :: PrivilegeTier) (p :: Phase). CanManageValidPaths t => Database t -> Set StorePath -> TenM p t ()

Context-aware mark paths as invalid

scanFileForReferences :: forall (t :: PrivilegeTier) (p :: Phase). CanScanReferences t => FilePath -> FilePath -> TenM p t (Set StorePath)

Context-aware file reference scanner

scanStoreForReferences :: forall (t :: PrivilegeTier) (p :: Phase). CanScanReferences t => FilePath -> TenM p t (Set StorePathReference)

Context-aware store reference scanner

vacuumReferenceDb :: forall (t :: PrivilegeTier) (p :: Phase). CanManageDatabase t => Database t -> TenM p t ()

Context-aware vacuum reference database

validateReferenceDb :: forall (t :: PrivilegeTier) (p :: Phase). CanManageDatabase t => Database t -> FilePath -> TenM p t (Int, Int)

Context-aware validate reference database

cleanupDanglingReferences :: forall (t :: PrivilegeTier) (p :: Phase). CanManageDatabase t => Database t -> TenM p t Int

Context-aware cleanup dangling references

data ReferenceEntry

A reference from one path to another

Constructors

ReferenceEntry

Fields

refFrom :: !StorePath

Path that contains the reference

refTo :: !StorePath

Path that is being referenced

data ReferenceStats

Statistics about references

Constructors

ReferenceStats

Fields

totalReferences :: !Int

Total number of references

uniqueReferrers :: !Int

Number of unique referrers

uniqueReferences :: !Int

Number of unique references

avgReferencesPerPath :: !Double

Average references per path

data StorePathReference

Reference from one store path to another

Constructors

StorePathReference

Fields

refReferrer :: !StorePath

refReference :: !StorePath