Ten.Daemon.Protocol

Documentation

compatibleVersions :: [ProtocolVersion]

List of compatible protocol versions

getDefaultSocketPath :: IO FilePath

data DaemonCapability

Daemon capabilities - mapped to privilege tiers

Constructors

StoreAccess

SandboxCreation

GarbageCollection

DerivationRegistration

DerivationBuild

StoreQuery

BuildQuery

DatabaseAccess

requestCapabilities :: DaemonRequest -> Set DaemonCapability

Determine capabilities required for a request

verifyCapabilities :: forall (t :: PrivilegeTier). SPrivilegeTier t -> Set DaemonCapability -> Either PrivilegeError ()

Check if a request has the necessary capabilities

checkPrivilegeRequirement :: forall (t :: PrivilegeTier). SPrivilegeTier t -> RequestPrivilege -> Either PrivilegeError ()

Check if a request can be performed with given privilege tier

data RequestPrivilege

Request privilege tagging

Constructors

PrivilegedRequest

UnprivilegedRequest

data ResponsePrivilege

Response privilege tagging

Constructors

PrivilegedResponse

UnprivilegedResponse

data PrivilegeRequirement

Privilege requirement for operations

Constructors

DaemonRequired

BuilderSufficient

data PrivilegeError

Privilege errors

Constructors

InsufficientPrivileges Text

InvalidCapability Text

AuthorizationError Text

data UserCredentials

User authentication credentials

Constructors

UserCredentials

Fields

username :: Text

token :: Text

requestedTier :: PrivilegeTier

data AuthRequestContent

Authentication request content

Constructors

AuthRequestContent

Fields

authVersion :: ProtocolVersion

authUser :: Text

authToken :: Text

authRequestedTier :: PrivilegeTier

data DaemonRequest

Daemon request types for protocol communication

Constructors

AuthRequest AuthRequestContent

BuildRequest Text (Maybe ByteString) BuildRequestInfo

EvalRequest Text (Maybe ByteString) BuildRequestInfo

BuildDerivationRequest Derivation BuildRequestInfo

BuildStatusRequest BuildId

CancelBuildRequest BuildId

QueryBuildOutputRequest BuildId

ListBuildsRequest (Maybe Int)

StoreAddRequest Text ByteString

StoreVerifyRequest StorePath

StorePathRequest Text ByteString

StoreListRequest

StoreReadRequest StorePath

StoreDerivationRequest ByteString

RetrieveDerivationRequest StorePath

QueryDerivationRequest Text Text (Maybe Int)

GetDerivationForOutputRequest Text

ListDerivationsRequest (Maybe Int)

GCRequest GCRequestParams

GCStatusRequest GCStatusRequestParams

AddGCRootRequest StorePath Text Bool

RemoveGCRootRequest Text

ListGCRootsRequest

PingRequest

ShutdownRequest

StatusRequest

ConfigRequest

DBQueryRequest DBQueryParams

DBExecuteRequest DBExecuteParams

DBTransactionBeginRequest TransactionMode

DBTransactionCommitRequest

DBTransactionRollbackRequest

DBGetSchemaVersionRequest

DBUpdateSchemaVersionRequest Int

RegisterDerivationFileRequest StorePath ByteString

RegisterDerivationWithOutputsRequest StorePath ByteString

GetReferencesToPathRequest StorePath

GetReferencesFromPathRequest StorePath

BulkRegisterReferencesRequest [(StorePath, StorePath)]

data DBQueryParams

Database query parameters

Constructors

DBQueryParams

Fields

dbQueryStatement :: Text

dbQueryParams :: [ByteString]

data DBExecuteParams

Database execute parameters

Constructors

DBExecuteParams

Fields

dbExecuteStatement :: Text

dbExecuteParams :: [ByteString]

data TransactionMode

Transaction modes

Constructors

ReadOnly
Read-only transaction

ReadWrite
Read-write transaction

Exclusive
Exclusive access

serializeDaemonRequest :: DaemonRequest -> (ByteString, Maybe ByteString)

Serialize a daemon request to ByteString

deserializeDaemonRequest :: ByteString -> Maybe ByteString -> Either Text DaemonRequest

Deserialize a daemon request from ByteString

serializeDaemonResponse :: DaemonResponse -> (ByteString, Maybe ByteString)

Serialize a daemon response to ByteString

createRequestFrame :: ByteString -> ByteString

Create a framed request for sending over the wire Format: [4-byte message length][message data]

parseRequestFrame :: ByteString -> Either Text (ByteString, ByteString)

Parse a framed request from the wire

createResponseFrame :: ByteString -> ByteString

Create a framed response for sending over the wire

parseResponseFrame :: ByteString -> Either Text (ByteString, ByteString)

Parse a framed response from the wire

sendDaemonRequest :: ProtocolHandle -> DaemonRequest -> IO (Either ProtocolError ByteString)

Send a daemon request through a protocol handle

receiveDaemonResponse :: Handle -> IO (Either ProtocolError ByteString)

Receive a daemon response from a handle

sendDaemonResponse :: ProtocolHandle -> DaemonResponse -> IO ()

Send a daemon response through a protocol handle

receiveDaemonRequest :: Handle -> IO (Either ProtocolError DaemonRequest)

Receive a daemon request from a handle

data ProtocolHandle

Protocol handle type for managing connections

Constructors

ProtocolHandle

Fields

protocolHandle :: Handle

Handle for all I/O operations

protocolLock :: MVar ()

For thread safety

protocolPrivilegeTier :: PrivilegeTier

Track privilege level of connection

createProtocolHandle :: Socket -> PrivilegeTier -> IO ProtocolHandle

Create a protocol handle from a socket

closeProtocolHandle :: ProtocolHandle -> IO ()

Close a protocol handle

withProtocolHandle :: Socket -> PrivilegeTier -> (ProtocolHandle -> IO a) -> IO a

Execute an action with a protocol handle and clean up after

requestToText :: DaemonRequest -> Text

Convert a request to human-readable text

responseToText :: DaemonResponse -> Text

Convert a response to human-readable text

data ProtocolError

Protocol errors

Constructors

ProtocolParseError Text

VersionMismatch ProtocolVersion ProtocolVersion

MessageTooLarge Word32

ConnectionClosed

AuthenticationFailed Text

OperationFailed Text

InvalidRequest Text

ProtocolInternalError Text

PrivilegeViolation Text

errorToText :: BuildError -> Text

Convert BuildError to Text

responseToResponseData :: Value -> Either Text DaemonResponse

Convert from Response payload to ResponseData

responsePrivilegeRequirement :: DaemonResponse -> ResponsePrivilege

Get privilege requirement for a response

parseBuildId :: Text -> Either Text BuildId

Parse a BuildId from Text

renderBuildId :: BuildId -> Text

Render a BuildId to Text

serializeQueryParams :: [ByteString] -> Text

Serialize query parameters

deserializeQueryResults :: ByteString -> Either Text a

Deserialize query results

getRowsAffected :: DaemonResponse -> Maybe Int64

Get rows affected from a response

Orphan instances

FromJSON Derivation

Instance details

Methods

parseJSON :: Value -> Parser Derivation

parseJSONList :: Value -> Parser [Derivation]

omittedField :: Maybe Derivation

ToJSON DaemonConfig

Instance details

Methods

toJSON :: DaemonConfig -> Value

toEncoding :: DaemonConfig -> Encoding

toJSONList :: [DaemonConfig] -> Value

toEncodingList :: [DaemonConfig] -> Encoding

omitField :: DaemonConfig -> Bool

ToJSON Derivation

Instance details

Methods

toJSON :: Derivation -> Value

toEncoding :: Derivation -> Encoding

toJSONList :: [Derivation] -> Value

toEncodingList :: [Derivation] -> Encoding

omitField :: Derivation -> Bool

ToJSON GCStats

Instance details

Methods

toJSON :: GCStats -> Value

toEncoding :: GCStats -> Encoding

toJSONList :: [GCStats] -> Value

toEncodingList :: [GCStats] -> Encoding

omitField :: GCStats -> Bool