Ten.Daemon.Server

Documentation

startServer :: Socket -> DaemonState 'Daemon -> DaemonConfig -> IO ServerControl

Start the server

stopServer :: ServerControl -> IO ()

Stop the server

acceptClients :: Socket -> ActiveClients -> TVar Bool -> DaemonState 'Daemon -> DaemonConfig -> TVar AuthDb -> TVar (Map SockAddr (Int, UTCTime)) -> Handle -> Handle -> TenM 'Build 'Daemon ()

Main loop to accept client connections with proper privilege evidence

handleClient :: Socket -> Handle -> ActiveClients -> DaemonState 'Daemon -> DaemonConfig -> TVar AuthDb -> TVar (Map SockAddr (Int, UTCTime)) -> Handle -> Handle -> TVar UTCTime -> TVar ClientState -> TVar Int -> TVar (Set Permission) -> TVar PrivilegeTier -> SockAddr -> TenM 'Build 'Daemon ()

Handle a client connection with proper privilege transitions

processRequest :: forall (t :: PrivilegeTier). StoreContentOps t => SPrivilegeTier t -> DaemonRequest -> DaemonState 'Daemon -> DaemonConfig -> Set Permission -> TenM 'Build t DaemonResponse

Process a client request with proper privilege tier

dispatchRequest :: forall (t :: PrivilegeTier). StoreContentOps t => SPrivilegeTier t -> DaemonRequest -> DaemonState 'Daemon -> DaemonConfig -> Set Permission -> TenM 'Build t DaemonResponse

Dispatch a request to the appropriate handler based on privilege tier

authenticateClient :: Handle -> TVar AuthDb -> TVar (Map SockAddr (Int, UTCTime)) -> SockAddr -> Handle -> IO (Either Text (UserId, AuthToken, Set Permission, PrivilegeTier))

Authenticate a client

handleBuildRequest :: forall (t :: PrivilegeTier). SPrivilegeTier t -> Text -> Maybe ByteString -> BuildRequestInfo -> DaemonState 'Daemon -> Set Permission -> TenM 'Build t DaemonResponse

Handler implementations

handleEvalRequest :: forall (t :: PrivilegeTier). SPrivilegeTier t -> Text -> Maybe ByteString -> BuildRequestInfo -> DaemonState 'Daemon -> Set Permission -> TenM 'Build t DaemonResponse

handleBuildDerivationRequest :: forall (t :: PrivilegeTier). SPrivilegeTier t -> Derivation -> BuildRequestInfo -> DaemonState 'Daemon -> Set Permission -> TenM 'Build t DaemonResponse

handleCancelBuildRequest :: forall (t :: PrivilegeTier). SPrivilegeTier t -> BuildId -> DaemonState 'Daemon -> Set Permission -> TenM 'Build t DaemonResponse

handleBuildStatusRequest :: forall (t :: PrivilegeTier). SPrivilegeTier t -> BuildId -> DaemonState 'Daemon -> Set Permission -> TenM 'Build t DaemonResponse

handleStoreRequest :: forall (t :: PrivilegeTier). StoreContentOps t => SPrivilegeTier t -> StoreCommand -> DaemonState 'Daemon -> DaemonConfig -> Set Permission -> TenM 'Build t DaemonResponse

handleGCRequest :: forall (t :: PrivilegeTier). SPrivilegeTier t -> Bool -> DaemonState 'Daemon -> DaemonConfig -> Set Permission -> TenM 'Build t DaemonResponse

handleStatusRequest :: forall (t :: PrivilegeTier). SPrivilegeTier t -> DaemonState 'Daemon -> DaemonConfig -> Set Permission -> TenM 'Build t DaemonResponse

handleConfigRequest :: forall (t :: PrivilegeTier). SPrivilegeTier t -> DaemonState 'Daemon -> DaemonConfig -> Set Permission -> TenM 'Build t DaemonResponse

handleShutdownRequest :: forall (t :: PrivilegeTier). SPrivilegeTier t -> DaemonState 'Daemon -> DaemonConfig -> Set Permission -> TenM 'Build t DaemonResponse

handleDerivationRequest :: forall (t :: PrivilegeTier). SPrivilegeTier t -> DerivationCommand -> DaemonState 'Daemon -> DaemonConfig -> Set Permission -> TenM 'Build t DaemonResponse

createServerSocket :: FilePath -> IO Socket

Create a server socket

closeServerSocket :: Socket -> FilePath -> IO ()

Close a server socket

data ClientInfo

Information about a connected client

Constructors

ClientInfo

Fields

ciThreadId :: ThreadId

Thread ID handling the client

ciSocket :: Socket

Client socket

ciHandle :: Handle

Socket handle for I/O

ciUserId :: UserId

Authenticated user ID

ciAuthToken :: AuthToken

Authentication token

ciConnectTime :: UTCTime

When the client connected

ciLastActivity :: TVar UTCTime

Last client activity time

ciState :: TVar ClientState

Current client state

ciAddress :: SockAddr

Client socket address

ciRequestCount :: TVar Int

Number of requests processed

ciPermissions :: TVar (Set Permission)

Client permissions

ciPrivilegeTier :: TVar PrivilegeTier

Assigned privilege tier

data ClientState

Client state tracking

Constructors

Authenticating
Client is in authentication process

Active
Client is actively connected

ShuttingDown
Client is being disconnected

type ActiveClients = TVar (Map ThreadId ClientInfo)

Collection of active client connections

addClient :: ActiveClients -> ClientInfo -> IO ()

Add a client to the active clients map

removeClient :: ActiveClients -> ThreadId -> IO ()

Remove a client from the active clients map

broadcastToClients :: ActiveClients -> ByteString -> IO ()

Broadcast a message to all clients

validateRequestPermissions :: forall (t :: PrivilegeTier). DaemonRequest -> Set Permission -> Handle -> SockAddr -> SPrivilegeTier t -> TenM 'Build t Bool

Validate permissions for a request with privilege tier context

logSecurityEvent :: ServerControl -> Text -> IO ()

Log a security event

data ServerControl

Server socket and control information

Constructors

ServerControl

Fields

scSocket :: Socket

Server listening socket

scThread :: ThreadId

Main server thread

scClients :: ActiveClients

Connected clients

scShutdown :: TVar Bool

Shutdown flag

scState :: DaemonState 'Daemon

Daemon state with phantom type

scConfig :: DaemonConfig

Daemon configuration

scAuthDb :: TVar AuthDb

Authentication database

scRateLimiter :: TVar (Map SockAddr (Int, UTCTime))

Rate limiting

scSecurityLog :: Handle

Security event log handle

scAccessLog :: Handle

Access log handle

scProcesses :: TVar (Map BuildId ProcessID)

Active build processes