Ten.Daemon.State

Documentation

data DaemonState (t :: PrivilegeTier)

Comprehensive daemon state with privilege phantom type parameter

Constructors

DaemonState

Fields

dsActiveBuilds :: TVar (Map BuildId ActiveBuild)

dsBuildQueue :: BuildQueue

dsCompletedBuilds :: TVar (Map BuildId (UTCTime, Either BuildError BuildResult))

dsFailedBuilds :: TVar (Map BuildId (UTCTime, BuildError))

dsPathLocks :: TVar (Map StorePath (TMVar ()))

dsReachablePaths :: TVar (Set StorePath)

dsKnownDerivations :: TVar (Map Text Derivation)

dsDerivationGraphs :: TVar (Map Text BuildGraph)

dsGCLock :: TMVar ()

dsGCLockOwner :: TVar (Maybe ProcessID)

dsGCLockPath :: FilePath

dsLastGC :: TVar (Maybe UTCTime)

dsGCStats :: TVar (UTCTime, Int, Int, Integer)

dsDaemonStats :: TVar DaemonStats

dsStartTime :: UTCTime

dsMaintenanceThread :: TVar (Maybe (Async ()))

dsStateFilePath :: FilePath

dsMaxCompletedBuilds :: Int

dsMaxFailedBuilds :: Int

dsMaxConcurrentBuilds :: Int

dsPrivilegeEvidence :: SPrivilegeTier t

data StateError

State errors

Constructors

StateBuildNotFound BuildId

StatePathNotFound StorePath

StateDerivationNotFound Text

StateIOError IOException

StateLockError Text

StateFormatError Text

StateVersionError Text

StateResourceError Text

StatePrivilegeError Text

initDaemonState :: forall (t :: PrivilegeTier). SPrivilegeTier t -> FilePath -> Int -> Int -> IO (DaemonState t)

Initialize a new daemon state with explicit privilege tier

loadStateFromFile :: forall (t :: PrivilegeTier). SPrivilegeTier t -> FilePath -> Int -> Int -> IO (DaemonState t)

Load daemon state from a file with proper privilege tier

saveStateToFile :: forall (t :: PrivilegeTier). (CanAccessStore t ~ 'True, CanModifyStore t ~ 'True) => DaemonState t -> IO ()

Save daemon state to a file - requires daemon privilege for file operations

populateState :: forall (t :: PrivilegeTier). DaemonState t -> Value -> IO ()

Populate state from loaded data (same implementation, but now takes a state with phantom type)

withDaemonState :: forall (t :: PrivilegeTier) a. SPrivilegeTier t -> (forall (s :: PrivilegeTier). SPrivilegeTier s -> DaemonState s -> IO a) -> DaemonState t -> IO a

Helper function to work with daemon state with a specific privilege tier

asDaemonState :: forall (t :: PrivilegeTier). DaemonState t -> IO (DaemonState 'Daemon)

Convert state to daemon privilege tier if possible (errors if not possible)

asBuilderState :: forall (t :: PrivilegeTier). DaemonState t -> DaemonState 'Builder

Convert state to builder privilege tier (always safe since we can drop privileges)

transitionStatePrivilege :: DaemonState 'Daemon -> DaemonState 'Builder

Transition state from one privilege tier to another (can only drop privileges)

data ActiveBuild

Active build information

Constructors

ActiveBuild

Fields

abBuildId :: BuildId

abDerivation :: Derivation

abOwner :: UserId

abStatus :: TVar BuildStatus

abStartTime :: UTCTime

abUpdateTime :: TVar UTCTime

abLogBuffer :: TVar Text

abResult :: TMVar (Either BuildError BuildResult)

abResourceUsage :: TVar (Map Text Double)

abDerivationChain :: TVar [Derivation]

abProcessId :: TVar (Maybe ProcessID)

abTimeout :: Maybe Int

abThread :: TVar (Maybe (Async ()))

abPrivilegeTier :: PrivilegeTier

data BuildQueue

Build queue

Constructors

BuildQueue

Fields

bqEntries :: TVar [BuildQueueEntry]

bqMaxPending :: Int

registerBuild :: forall (t :: PrivilegeTier). DaemonState t -> Derivation -> UserId -> Int -> Maybe Int -> IO BuildId

Register a new build - available for both privilege tiers

unregisterBuild :: forall (t :: PrivilegeTier). DaemonState t -> BuildId -> IO ()

Unregister a build

updateBuildStatus :: forall (t :: PrivilegeTier). DaemonState t -> BuildId -> BuildStatus -> IO ()

Update a build's status

appendBuildLog :: forall (t :: PrivilegeTier). DaemonState t -> BuildId -> Text -> IO ()

Append to a build's log

getBuildStatus :: forall (t :: PrivilegeTier). DaemonState t -> BuildId -> IO BuildStatus

Get a build's status

getBuildLog :: forall (t :: PrivilegeTier). DaemonState t -> BuildId -> IO Text

Get a build's log

listActiveBuilds :: forall (t :: PrivilegeTier). DaemonState t -> IO [(BuildId, Derivation, BuildStatus, UTCTime, Double)]

List active builds

listQueuedBuilds :: forall (t :: PrivilegeTier). DaemonState t -> IO [(BuildId, Derivation, Int, UTCTime)]

List queued builds

registerReturnedDerivation :: forall (t :: PrivilegeTier). DaemonState t -> BuildId -> Derivation -> IO Bool

Register a returned derivation for a build

getDerivationChain :: forall (t :: PrivilegeTier). DaemonState t -> BuildId -> IO [Derivation]

Get the derivation chain for a build

isDerivationInChain :: forall (t :: PrivilegeTier). DaemonState t -> BuildId -> Derivation -> IO Bool

Check if a derivation is in a build's chain

acquirePathLock :: forall (t :: PrivilegeTier). DaemonState t -> StorePath -> IO ()

Acquire a lock for a store path

releasePathLock :: forall (t :: PrivilegeTier). DaemonState t -> StorePath -> IO ()

Release a lock for a store path

withPathLock :: forall (t :: PrivilegeTier) a. DaemonState t -> StorePath -> IO a -> IO a

Execute an action with a locked store path

markPathAsReachable :: forall (t :: PrivilegeTier). CanModifyStore t ~ 'True => DaemonState t -> StorePath -> IO ()

Mark a path as reachable - requires CanModifyStore privilege

isPathReachable :: forall (t :: PrivilegeTier). DaemonState t -> StorePath -> IO Bool

Check if a path is reachable

getReachablePaths :: forall (t :: PrivilegeTier). DaemonState t -> IO (Set StorePath)

Get all reachable paths

registerDerivationGraph :: forall (t :: PrivilegeTier). DaemonState t -> Derivation -> BuildGraph -> IO ()

Register a derivation's build graph

getDerivationGraph :: forall (t :: PrivilegeTier). DaemonState t -> Derivation -> IO (Maybe BuildGraph)

Get a derivation's build graph

getTransitiveDependencies :: forall (t :: PrivilegeTier). DaemonState t -> Derivation -> IO (Set Text)

Get transitive dependencies of a derivation

acquireGCLock :: forall (t :: PrivilegeTier). CanModifyStore t ~ 'True => DaemonState t -> IO ()

Acquire the GC lock - requires CanModifyStore privilege

releaseGCLock :: forall (t :: PrivilegeTier). CanModifyStore t ~ 'True => DaemonState t -> IO ()

Release the GC lock - requires CanModifyStore privilege

withGCLock :: forall (t :: PrivilegeTier) a. CanModifyStore t ~ 'True => DaemonState t -> IO a -> IO a

Execute an action with the GC lock - requires CanModifyStore privilege

checkGCLock :: forall (t :: PrivilegeTier). DaemonState t -> STM Bool

Check if GC lock is available (can be acquired)

tryAcquireGCLock :: forall (t :: PrivilegeTier). CanModifyStore t ~ 'True => DaemonState t -> IO Bool

Try to acquire the GC lock without throwing

data DaemonStats

System statistics

Constructors

DaemonStats

Fields

statsBuildsTotalCompleted :: Int

statsBuildsTotalFailed :: Int

statsCPUUsage :: Double

statsMemoryUsage :: Integer

statsStoreSize :: Integer

statsStoreEntries :: Int

statsSystemLoad :: [Double]

statsLastUpdated :: UTCTime

updateSystemStats :: forall (t :: PrivilegeTier). DaemonState t -> IO ()

Update system statistics

getDaemonStats :: forall (t :: PrivilegeTier). DaemonState t -> IO DaemonStats

Get daemon statistics

hasPendingBuilds :: forall (t :: PrivilegeTier). DaemonState t -> IO Bool

Check if there are pending builds

hasActiveBuild :: forall (t :: PrivilegeTier). DaemonState t -> BuildId -> IO Bool

Check if a specific build is active

getNextBuildToSchedule :: forall (t :: PrivilegeTier). DaemonState t -> IO (Maybe BuildQueueEntry)

Get the next build to schedule

hasBuildCapacity :: forall (t :: PrivilegeTier). DaemonState t -> IO Bool

Check if there is capacity for more builds

pruneCompletedBuilds :: forall (t :: PrivilegeTier). DaemonState t -> IO ()

Prune old completed builds to stay within limits

cleanupStaleBuilds :: forall (t :: PrivilegeTier). DaemonState t -> IO ()

Clean up stale builds (e.g., timed out or crashed)

scheduledMaintenance :: forall (t :: PrivilegeTier). (CanModifyStore t ~ 'True, CanAccessStore t ~ 'True) => DaemonState t -> IO ()

Run scheduled maintenance tasks - requires CanModifyStore privilege

atomicallyState :: STM a -> IO a

Execute an STM transaction in the context of daemon state

withState :: forall (t :: PrivilegeTier) a. DaemonState t -> (DaemonState t -> IO a) -> IO a

Execute an action with daemon state

modifyState :: forall (t :: PrivilegeTier) a. DaemonState t -> (DaemonState t -> IO (DaemonState t, a)) -> IO a

Modify daemon state

readState :: forall (t :: PrivilegeTier) a. DaemonState t -> (DaemonState t -> IO a) -> IO a

Read daemon state

captureStatus :: forall (t :: PrivilegeTier). DaemonState t -> BuildId -> IO (BuildStatus, Double, UTCTime, UTCTime)

Capture build status information

calculateBuildProgress :: ActiveBuild -> IO Double

Calculate build progress